+++ src/training/steps/backtesting/nas_tas/__init__.py
@@ -6,7 +6,7 @@ and model training systems to provide complete validation capabilities.
 """
 
 from .backtesting_engine import BacktestingEngine, BacktestingConfig
-from .walk_forward_analyzer import WalkForwardAnalyzer, WalkForwardConfig
+from src.utils.nas_tas.walk_forward_analyzer import WalkForwardAnalyzer, WalkForwardConfig
 from src.utils.nas_tas.performance_attribution import PerformanceAttributor, AttributionConfig
 from .scenario_tester import ScenarioTester, ScenarioConfig
 from .validation_orchestrator import ValidationOrchestrator, ValidationConfig
