+++ src/training/steps/market_analysis/tas_regime/backtesting/__init__.py
@@ -12,7 +12,7 @@ Comprehensive backtesting framework for tree architecture search including:
 """
 
 from .backtesting_engine import BacktestingEngine, BacktestingConfig, BacktestingResult
-from .walk_forward_analysis import WalkForwardAnalyzer, WalkForwardConfig
+from src.utils.nas_tas.walk_forward_analyzer import WalkForwardAnalyzer, WalkForwardConfig
 from src.utils.nas_tas.performance_attribution import PerformanceAttributor, AttributionConfig
 from .risk_analysis import RiskAnalyzer, RiskConfig
 from .scenario_testing import ScenarioTester, ScenarioConfig
