+++ src/utils/ml_common/optimization/tas/backtesting/BACKTESTING_IMPLEMENTATION_SUMMARY.md
@@ -17,7 +17,7 @@ The comprehensive backtesting framework for TAS has been successfully implemente
   - Results export and storage
 
 ### 2. **Walk-Forward Analysis** ✅
-- **File**: `walk_forward_analysis.py`
+- **File**: `src/utils/nas_tas/walk_forward_analyzer.py` (unified)
 - **Features**:
   - Rolling window analysis
   - Expanding window analysis
@@ -96,7 +96,7 @@ The comprehensive backtesting framework for TAS has been successfully implemente
 backtesting/
 ├── __init__.py                          # Package initialization
 ├── backtesting_engine.py               # Core backtesting engine
-├── walk_forward_analysis.py            # Walk-forward analysis
+├── (unified) → src/utils/nas_tas/walk_forward_analyzer.py  # Walk-forward analysis
 ├── performance_attribution.py          # Performance attribution
 ├── risk_analysis.py                    # Risk analysis
 ├── scenario_testing.py                 # Scenario testing
