+++ src/utils/ml_common/optimization/tas/backtesting/examples/backtesting_example.py
@@ -19,7 +19,7 @@ warnings.filterwarnings('ignore')
 
 # Import TAS backtesting components
 from ..backtesting_engine import BacktestingEngine, BacktestingConfig, BacktestingResult
-from ..walk_forward_analysis import WalkForwardAnalyzer, WalkForwardConfig, WalkForwardResult
+from src.utils.nas_tas.walk_forward_analyzer import WalkForwardAnalyzer, WalkForwardConfig, WalkForwardResult
 from src.utils.nas_tas.performance_attribution import PerformanceAttributor, AttributionConfig, AttributionResult
 from ..risk_analysis import RiskAnalyzer, RiskConfig, RiskResult
 from ..scenario_testing import ScenarioTester, ScenarioConfig, ScenarioResult
